luminex_sdk/requirements.txt
pygame==2.6.0
pyglet==2.0.0
PyOpenGL==3.1.5
